news preview refactor/dark mode support
