+++ src/app/features/games/pages/games.component.html
@@ -9,7 +9,9 @@
             : ''
         "
       ></dd-card-header>
-      <dd-news-container [category]="urlValue"></dd-news-container>
+      <dd-card [title]="'News & Updates'">
+        <dd-news-preview-list class="col-span-2"></dd-news-preview-list
+      ></dd-card>
     </div>
   </div>
 </div>
