--- src/app/shared/components/card-header/card-header.component.html
@@ -1,5 +1,0 @@
-<h3
-  class="flex justify-center font-semibold bg-byzantium rounded-t p-2 border-b border-zinc-800"
->
-  {{ title }}
-</h3>
