+++ src/app/shared/components/card/card-header/card-header.component.html
@@ -1,0 +1,5 @@
+<h3
+  class="flex justify-center font-semibold bg-byzantium rounded-t p-2 border-b border-zinc-800 text-zinc-200"
+>
+  {{ title }}
+</h3>
