--- src/app/shared/components/news-container/news-container.component.html
@@ -1,12 +1,0 @@
-<div class="bg-light_secondary dark:bg-dark_secondary rounded w-full">
-  <dd-card-header [title]="'News & Updates'" class="text-sm"></dd-card-header>
-  <div class="flex flex-col gap-3 items-center p-4">
-    <ng-container *ngFor="let group of posts">
-      <div class="w-full flex">{{ group[0] }}</div>
-      <dd-news-card *ngFor="let post of group[1]" [post]="post"></dd-news-card>
-    </ng-container>
-  </div>
-  <ng-template #elseBlock>
-    <dd-spinner></dd-spinner>
-  </ng-template>
-</div>
