+++ src/app/shared/components/news-preview/news-preview.component.html
@@ -26,7 +26,7 @@
     <div class="sm:flex sm:items-end sm:justify-end">
       <a
         [routerLink]="['/', 'news', post.slug.current]"
-        class="block bg-byzantium px-5 py-3 text-center text-xs font-semibold uppercase transition hover:bg-old_mauve rounded-br"
+        class="block bg-byzantium px-5 py-3 text-center text-xs font-semibold uppercase transition hover:bg-old_mauve rounded-br text-zinc-200"
       >
         View Full Post
       </a>
