+++ src/app/shared/shared.module.ts
@@ -5,10 +5,8 @@ import { DecimetersToPoundsPipe } from './pipes/decimeters-to-pounds/decimeters-
 import { FormsModule } from '@angular/forms';
 import { TypeBoxComponent } from './components/type-box/type-box.component';
 import { HeaderDropdownItemComponent } from './components/header-dropdown-item/header-dropdown-item.component';
-import { PostCardComponent } from './components/news-card/news-card.component';
-import { CardHeaderComponent } from './components/card-header/card-header.component';
+import { CardHeaderComponent } from './components/card/card-header/card-header.component';
 import { RouterModule } from '@angular/router';
-import { NewsContainerComponent } from './components/news-container/news-container.component';
 import { SpinnerComponent } from './components/spinner/spinner.component';
 import { DropdownLinkComponent } from './components/dropdown-link/dropdown-link.component';
 import { PokedexSelectDropdownComponent } from './components/pokedex-select-dropdown/pokedex-select-dropdown.component';
@@ -16,20 +14,22 @@ import { NgIconsModule } from '@ng-icons/core';
 import { heroChevronDownSolid } from '@ng-icons/heroicons/solid';
 import { CardComponent } from './components/card/card.component';
 import { ButtonComponent } from './components/button/button.component';
+import { NewsPreviewComponent } from './components/news-preview/news-preview.component';
+import { NewsPreviewListComponent } from './components/news-preview-list/news-preview-list.component';
 
 @NgModule({
   declarations: [
     DecimetersToFeetAndInchesPipe,
     DecimetersToPoundsPipe,
     TypeBoxComponent,
-    PostCardComponent,
+    NewsPreviewComponent,
     DropdownLinkComponent,
     PokedexSelectDropdownComponent,
     HeaderDropdownItemComponent,
     ButtonComponent,
     CardComponent,
     CardHeaderComponent,
-    NewsContainerComponent,
+    NewsPreviewListComponent,
     SpinnerComponent,
   ],
   imports: [
@@ -47,11 +47,11 @@ import { ButtonComponent } from './components/button/button.component';
     HeaderDropdownItemComponent,
     DropdownLinkComponent,
     PokedexSelectDropdownComponent,
-    PostCardComponent,
+    NewsPreviewComponent,
     ButtonComponent,
     CardComponent,
     CardHeaderComponent,
-    NewsContainerComponent,
+    NewsPreviewListComponent,
     SpinnerComponent,
   ],
 })
